+++ cli/src/main.rs
@@ -3,6 +3,7 @@ use crossterm::style::Stylize;
 mod commands;
 mod fs;
 mod package_managers;
+mod parsers;
 
 fn main() {
     if let Err(error) = commands::tui::run_command() {